debug

+++ src/prune.py
@@ -554,7 +554,7 @@ class Pruner:
                     # weight = param.data
                     weight_dev = param.device
                     # contr_mask = (np.ones(weight.shape) * coef).astype("float32")
-                    param.data = torch.from_numpy(weight * control_weights).to(weight_dev)
+                    param.data = torch.from_numpy(weight * control_weights).to(weight_dev).astype("float32")
                     # new_weights = torch.mul(weight, control_weights)
                     # param.data = new_weights.to(weight_dev)
                     break
